// ClientApp/src/app/book-management/models/BookBindingInfo.ts
export interface BookBindingInfo {
    bookBindingInfoId:number;
    bookInformationId: number;
    accessionNo:string;
    pressName: string;
    baseSchoolNameId:number;
    pressNumber: number;
    date:Date;
    pressEmail: string;
    pressAddress: string;
    senderOfficer: string;
    officeDispiseNumber: string;
    receivedOfficerName: string;
    isActive: boolean;
    bookTitleEnglish:string;
    bookTitleBangla:string;
    category:string;
}